requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
gunicorn==25.0.2
packaging==26.0
sqlparse==0.5.5
whitenoise==6.11.0
yt-dlp==2026.2.4
google-genai
pydantic
requests
python-dotenv
celery
redis
dj-database-url
psycopg2-binary
ddgs
dj-database-url==3.1.2